Improved: moved the implementation of ICloverGoConnectorListener to a separate file (#2ft9uc)
