+++ src/android/CloverGo.java
@@ -1,4 +1,4 @@
-package cordova.plugin.clovergo;
+package co.hotwax.clovergo;
 
 import android.widget.Toast;
 import android.util.Log;
@@ -550,248 +550,5 @@ public class CloverGo extends CordovaPlugin {
         PluginResult result = new PluginResult(PluginResult.Status.ERROR, errorMsg);
         result.setKeepCallback(isKeepCallBack);
         rootCallbackContext.sendPluginResult(result);
-    }
-}
-
-// TODO Use a separate file for it
-/**
- * Adapter that implements only the required methods
- */
-
-class GoConnectorListener implements ICloverGoConnectorListener {
-
-    @Override
-    public void onDeviceActivityStart(CloverDeviceEvent deviceEvent) {
-
-    }
-
-    @Override
-    public void onDeviceActivityEnd(CloverDeviceEvent deviceEvent) {
-
-    }
-
-    @Override
-    public void onDeviceError(CloverDeviceErrorEvent deviceErrorEvent) {
-
-    }
-
-    @Override
-    public void onPreAuthResponse(PreAuthResponse response) {
-
-    }
-
-    @Override
-    public void onAuthResponse(AuthResponse response) {
-
-    }
-
-    @Override
-    public void onTipAdjustAuthResponse(TipAdjustAuthResponse response) {
-
-    }
-
-    @Override
-    public void onCapturePreAuthResponse(CapturePreAuthResponse response) {
-
-    }
-
-    @Override
-    public void onVerifySignatureRequest(VerifySignatureRequest request) {
-
-    }
-
-    @Override
-    public void onConfirmPaymentRequest(ConfirmPaymentRequest request) {
-
-    }
-
-    @Override
-    public void onCloseoutResponse(CloseoutResponse response) {
-
-    }
-
-    @Override
-    public void onSaleResponse(SaleResponse response) {
-
-    }
-
-    @Override
-    public void onManualRefundResponse(ManualRefundResponse response) {
-
-    }
-
-    @Override
-    public void onRefundPaymentResponse(RefundPaymentResponse response) {
-
-    }
-
-    @Override
-    public void onTipAdded(TipAddedMessage message) {
-
-    }
-
-    @Override
-    public void onVoidPaymentResponse(VoidPaymentResponse response) {
-
-    }
-
-    @Override
-    public void onDeviceDisconnected() {
-
-    }
-
-    @Override
-    public void onDeviceConnected() {
-
-    }
-
-    @Override
-    public void onDeviceReady(MerchantInfo merchantInfo) {
-
-    }
-
-    @Override
-    public void onVaultCardResponse(VaultCardResponse response) {
-
-    }
-
-    @Override
-    public void onPrintJobStatusResponse(PrintJobStatusResponse response) {
-
-    }
-
-    @Override
-    public void onRetrievePrintersResponse(RetrievePrintersResponse response) {
-
-    }
-
-    @Override
-    public void onPrintManualRefundReceipt(PrintManualRefundReceiptMessage message) {
-
-    }
-
-    @Override
-    public void onPrintManualRefundDeclineReceipt(PrintManualRefundDeclineReceiptMessage message) {
-
-    }
-
-    @Override
-    public void onPrintPaymentReceipt(PrintPaymentReceiptMessage message) {
-
-    }
-
-    @Override
-    public void onPrintPaymentDeclineReceipt(PrintPaymentDeclineReceiptMessage message) {
-
-    }
-
-    @Override
-    public void onPrintPaymentMerchantCopyReceipt(PrintPaymentMerchantCopyReceiptMessage message) {
-
-    }
-
-    @Override
-    public void onPrintRefundPaymentReceipt(PrintRefundPaymentReceiptMessage message) {
-
-    }
-
-    @Override
-    public void onRetrievePendingPaymentsResponse(RetrievePendingPaymentsResponse response) {
-
-    }
-
-    @Override
-    public void onReadCardDataResponse(ReadCardDataResponse response) {
-
-    }
-
-    @Override
-    public void onMessageFromActivity(MessageFromActivity message) {
-
-    }
-
-    @Override
-    public void onCustomActivityResponse(CustomActivityResponse response) {
-
-    }
-
-    @Override
-    public void onRetrieveDeviceStatusResponse(RetrieveDeviceStatusResponse response) {
-
-    }
-
-    @Override
-    public void onResetDeviceResponse(ResetDeviceResponse response) {
-
-    }
-
-    @Override
-    public void onRetrievePaymentResponse(RetrievePaymentResponse response) {
-
-    }
-
-    @Override
-    public void onDeviceDiscovered(ReaderInfo readerInfo) {
-
-    }
-
-    @Override
-    public void onDeviceDisconnected(ReaderInfo readerInfo) {
-
-    }
-
-    @Override
-    public void onAidMatch(List<CardApplicationIdentifier> applicationIdentifierList, AidSelection aidSelection) {
-
-    }
-
-    @Override
-    public void onPaymentTypeRequired(int cardEntryMethods, List<ReaderInfo> connectedReaders, PaymentTypeSelection paymentTypeSelection) {
-
-    }
-
-    @Override
-    public void onManualCardEntryRequired(CloverGoConstants.TransactionType transactionType, BaseRequest baseRequest, ICloverGoConnector.GoPaymentType goPaymentType, ReaderInfo.ReaderType readerType, boolean allowDuplicate, ManualCardEntry manualCardEntry) {
-
-    }
-
-    @Override
-    public void notifyOnProgressDialog(String title, String message, boolean isCancelable) {
-
-    }
-
-    @Override
-    public void onCloverGoDeviceActivity(CloverDeviceEvent deviceEvent) {
-
-    }
-
-    @Override
-    public void onGetMerchantInfo() {
-
-    }
-
-    @Override
-    public void onGetMerchantInfoResponse(MerchantInfo merchantInfo) {
-
-    }
-
-    @Override
-    public void onSignatureRequired(Payment payment, SignatureCapture signatureCapture) {
-
-    }
-
-    @Override
-    public void onSendReceipt(Order order, SendReceipt sendReceipt) {
-
-    }
-
-    @Override
-    public void onDisplayMessage(String message) {
-
-    }
-
-    @Override
-    public void onVoidPayment(Payment payment, String reason) {
-
     }
 }
